Authentication flow transferred to AuthenticationActivity
