--- starter/app/src/main/java/com/udacity/project4/authentication/LoginFragment.kt
@@ -1,138 +1,0 @@
-/*
- * Copyright 2019, The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.udacity.project4.authentication
-
-import android.app.Activity
-import android.content.Intent
-import android.os.Bundle
-import android.util.Log
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.databinding.DataBindingUtil
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
-import androidx.navigation.fragment.findNavController
-import com.firebase.ui.auth.AuthUI
-import com.firebase.ui.auth.IdpResponse
-import com.google.firebase.auth.FirebaseAuth
-import com.udacity.project4.R
-import com.udacity.project4.databinding.FragmentLoginBinding
-
-class LoginFragment : Fragment() {
-
-    companion object {
-        const val TAG = "MainFragment"
-        const val SIGN_IN_RESULT_CODE = 1001
-    }
-
-    // Get a reference to the ViewModel scoped to this Fragment
-    private val viewModel by viewModels<LoginViewModel>()
-    private lateinit var binding: FragmentLoginBinding
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
-    ): View {
-        binding = DataBindingUtil.inflate(inflater, R.layout.fragment_login, container, false)
-        binding.authButton.text = getString(R.string.login_button_text)
-        return binding.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        observeAuthenticationState()
-
-        binding.authButton.setOnClickListener {
-            launchSignInFlow()
-        }
-    }
-
-    /**
-     * Listen to the result of the sign in process by filter for when
-     * SIGN_IN_REQUEST_CODE is passed back. Start by having log statements to know
-     * whether the user has signed in successfully
-     */
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (requestCode == SIGN_IN_RESULT_CODE) {
-            val response = IdpResponse.fromResultIntent(data)
-            if (resultCode == Activity.RESULT_OK) {
-                // User successfully signed in
-                Log.i(TAG, "Successfully signed in user " +
-                    "${FirebaseAuth.getInstance().currentUser?.displayName}!"
-                )
-                val action = LoginFragmentDirections.toReminder()
-                findNavController().navigate(action)
-            } else {
-                // Sign in failed. If response is null the user canceled the
-                // sign-in flow using the back button. Otherwise check
-                // response.getError().getErrorCode() and handle the error.
-                Log.i(TAG, "Sign in unsuccessful ${response?.error?.errorCode}")
-            }
-        }
-    }
-
-    /**
-     * Observes the authentication state and changes the UI accordingly.
-     * If there is a logged in user: (1) show a logout button and (2) display their name.
-     * If there is no logged in user: show a login button
-     */
-    private fun observeAuthenticationState() {
-        viewModel.authenticationState.observe(viewLifecycleOwner) { authenticationState ->
-            // in LoginViewModel and change the UI accordingly.
-            when (authenticationState) {
-                LoginViewModel.AuthenticationState.AUTHENTICATED -> { logoutFromFirebase() }
-                else -> { loginOnFirebase() }
-            }
-        }
-    }
-
-    private fun loginOnFirebase() {
-        binding.authButton.text = getString(R.string.login_button_text)
-        binding.authButton.setOnClickListener {
-            launchSignInFlow()
-        }
-    }
-
-    private fun logoutFromFirebase() {
-        binding.authButton.text = getString(R.string.logout_button_text)
-        binding.authButton.setOnClickListener {
-            AuthUI.getInstance().signOut(requireContext())
-        }
-    }
-
-    private fun launchSignInFlow() {
-        // Give users the option to sign in / register with their email or Google account.
-        // If users choose to register with their email,
-        // they will need to create a password as well.
-        val providers = arrayListOf(
-            AuthUI.IdpConfig.EmailBuilder().build(), AuthUI.IdpConfig.GoogleBuilder().build()
-            // This is where you can provide more ways for users to register and
-            // sign in.
-        )
-
-        // Create and launch sign-in intent.
-        // We listen to the response of this activity with the SIGN_IN_REQUEST_CODE
-        startActivityForResult(
-            AuthUI.getInstance()
-                .createSignInIntentBuilder()
-                .setAvailableProviders(providers)
-                .build(),
-            SIGN_IN_RESULT_CODE
-        )
-    }
-}
