+++ starter/app/src/main/res/values/dimens.xml
@@ -10,7 +10,7 @@
     <dimen name="padding_large">20dp</dimen>
     <dimen name="margin_normal">16dp</dimen>
 
-    <dimen name="text_margin">20dp</dimen>
+    <dimen name="text_margin">15dp</dimen>
     <dimen name="text_padding">10dp</dimen>
-    <dimen name="button_text_size">20sp</dimen>
+    <dimen name="button_text_size">16sp</dimen>
 </resources>
